pasha-bot/requirements.txt
python-telegram-bot
python-dotenv
requests
pytest
google-generativeai
httpx
apscheduler
aiosqlite